--- apps/web/modules/custom-header/runtime/components/MegaMenu/types.ts
@@ -1,5 +1,0 @@
-import type { CategoryTreeItem } from '@plentymarkets/shop-api';
-
-export type MegaMenuProps = {
-  categories: CategoryTreeItem[];
-};
