--- apps/web/modules/custom-header/runtime/layouts/default.vue
@@ -1,34 +1,0 @@
-<template>
-  <div>
-    <ModuleHeader />
-    <NarrowContainer v-if="breadcrumbs?.length" class="p-4 md:px-0">
-      <LazyUiBreadcrumbs :breadcrumbs="breadcrumbs" />
-    </NarrowContainer>
-    <main>
-      <slot />
-    </main>
-    <UiNavbarBottom v-if="viewport.isLessThan('lg')" />
-    <Cookiebar />
-    <PreviewMode />
-    <ClientOnly>
-      <FooterBlock v-if="!route.meta.isBlockified" />
-    </ClientOnly>
-    <QuickCheckout v-if="isOpen" :product="product" />
-  </div>
-</template>
-
-<script setup lang="ts">
-import type { DefaultLayoutProps } from '~/layouts/types';
-import FooterBlock from '~/components/blocks/Footer/Footer.vue';
-
-defineOptions({ name: 'DefaultLayout' });
-
-defineProps<DefaultLayoutProps>();
-
-const { setLogoMeta } = useStructuredData();
-const { isOpen, product } = useQuickCheckout();
-const viewport = useViewport();
-const route = useRoute();
-
-setLogoMeta();
-</script>
