+++ apps/web/modules/nk-override/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "@plentymarkets/pwa-module-boilerplate",
+  "name": "nk-override",
   "version": "1.0.0",
   "license": "MIT",
   "type": "module",
