+++ apps/web/modules/nk-override/types/tailwind.ts
