--- apps/web/modules/pwa-module-boilerplate-main/.github/CODEOWNERS
@@ -1,1 +1,0 @@
-*   @plentymarkets/plentyshop
