--- apps/web/modules/pwa-module-boilerplate-main/.gitignore
@@ -1,6 +1,0 @@
-node_modules
-dist
-.yarn/cache
-.yarn/install-state.gz
-.nuxt
-.idea
