--- apps/web/modules/pwa-module-boilerplate-main/.prettierignore
@@ -1,4 +1,0 @@
-node_modules
-.nuxt
-.yarn
-dist
