--- apps/web/modules/pwa-module-boilerplate-main/.prettierrc
@@ -1,10 +1,0 @@
-{
-  "endOfLine": "lf",
-  "semi": true,
-  "singleQuote": true,
-  "arrowParens": "always",
-  "tabWidth": 2,
-  "trailingComma": "all",
-  "printWidth": 120,
-  "jsxSingleQuote": false
-}
