--- apps/web/modules/pwa-module-boilerplate-main/README.md
@@ -1,8 +1,0 @@
-<div align="center">
-    <a href="https://www.plentymarkets.com/"><img src="https://cdn01.plentymarkets.com/avw8j9fg70hi/frontend/plentyShop_LTS/plentysystems-logo.jpg" alt="plentysystems logo" height="160px" /></a>
-    <h1>PlentyONE Shop Module Boilerplate</h1>
-</div>
-
-## Work in progress
-
-This project is a boilerplate for a PlentyONE Shop Module. It is a work in progress and is not yet ready for production use.
