--- apps/web/modules/pwa-module-boilerplate-main/src/index.ts
@@ -1,1 +1,0 @@
-export { default } from './module';
