--- apps/web/modules/pwa-module-boilerplate-main/src/runtime/lang/de.json
@@ -1,5 +1,0 @@
-{
-  "moduleBoilerplate": {
-    "button": "Drücken um hochzuzählen"
-  }
-}
