--- apps/web/modules/pwa-module-boilerplate-main/src/runtime/lang/en.json
@@ -1,5 +1,0 @@
-{
-  "moduleBoilerplate": {
-    "button": "Press to count up"
-  }
-}
