--- apps/web/modules/pwa-module-boilerplate-main/src/runtime/layouts/checkout.vue
@@ -1,7 +1,0 @@
-<template>
-  <div>
-    <main class="bg-red-200">
-      <slot />
-    </main>
-  </div>
-</template>
