--- apps/web/modules/pwa-module-boilerplate-main/src/runtime/middleware/example-middleware.ts
@@ -1,5 +1,0 @@
-import { defineNuxtRouteMiddleware } from 'nuxt/app';
-
-export default defineNuxtRouteMiddleware(async (to, from) => {
-  // This middleware will be called for every route change
-});
