--- apps/web/modules/pwa-module-boilerplate-main/src/runtime/pages/index.vue
@@ -1,5 +1,0 @@
-<template>
-  <div class="bg-amber-300">Override index.vue</div>
-</template>
-
-<script setup lang="ts"></script>
