--- apps/web/modules/pwa-module-boilerplate-main/src/runtime/plugins/registerComponents.ts
@@ -1,7 +1,0 @@
-import { defineNuxtPlugin } from 'nuxt/app';
-
-export default defineNuxtPlugin((nuxtApp) => {
-  if (import.meta.server) {
-    console.log('registerComponents.ts');
-  }
-});
