--- apps/web/modules/pwa-module-boilerplate-main/src/runtime/plugins/registerCookies.ts
@@ -1,22 +1,0 @@
-import { defineNuxtPlugin, useRuntimeConfig } from 'nuxt/app';
-
-export default defineNuxtPlugin((nuxtApp) => {
-  if (import.meta.server) {
-    const runtimeConfig = useRuntimeConfig();
-    const initialCookies = runtimeConfig.public?.cookieGroups;
-
-    if (initialCookies && initialCookies.groups) {
-      initialCookies.groups.forEach((cookieGroup) => {
-        if (cookieGroup.name === 'CookieBar.functional.label') {
-          cookieGroup.cookies.push({
-            name: 'paypalModule.paypal',
-            Provider: 'CookieBar.functional.cookies.payPal.provider',
-            Status: 'CookieBar.functional.cookies.payPal.status',
-            PrivacyPolicy: '/PrivacyPolicy',
-            Lifespan: 'Session',
-          });
-        }
-      });
-    }
-  }
-});
